Fixed and optimised add-book form layout in order to have a neat look and an intuitive form
Prior to this commit, It came to attention by one collaborator that one
of the fields in the form was not properly visible which is the select
genre. Modified the field and made sure it was visible and got rid of
redundant whites and closed in the form to have a neat look.